requirements.txt
streamlit==1.37.1
google-generativeai>=0.7.2
protobuf<5
pypdf>=4.2.0
